Add files via upload
These two files are an alternative of PesVres.sql. This schema’s Question table has a VARCHAR(3) primary key instead of INT AUTO_INCREMENT. The schema.sql file handles the infrastructure of our query database and data.sql is dedicated for the aggregation of all questions and choices to be inserted into the tables.